services/api/pyproject.toml
[tool.poetry]
name = "ssi-api"
version = "1.0.0"
description = "SSI HTTP production reference"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-jose = "^3.3.0"
jwcrypto = "^1.5.6"
cryptography = "^43.0.0"
pydantic = "^2.8.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.34"
redis = "^5.0.7"
python-slugify = "^8.0.4"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
structlog = "^24.1.0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-exporter-otlp = "^1.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
ruff = "^0.6.4"
